Fix H265/HEVC support
The relevant notes from the Chromium source are:

https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/mediarecorder/media_recorder_handler.cc;l=384-396;drc=3d0711d78cbda790954e547276858840c6dd6732

and

https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/mediarecorder/media_recorder_handler.cc;l=408-415;drc=3d0711d78cbda790954e547276858840c6dd6732

`hvc1` is the only accepted codec string for H265/HEVC and
additionally requires the .mp4 container format instead of .webm.

Thus move both container format and codec to the INPUT value and,
while on it, switch h264 to .mp4 as well - which then requires to
use the preferred codec name avc1.

This was only tested with Chromium, with the `enable_hevc_parser_and_hw_decoder`
build flag and `MediaRecorderHEVCSupport` runtime flag enabled, in an
effort to improve hardware de- and encoding on Linux.

According to the linked comments it should also work on Safari,
Firefox doesn't seem to implement HEVC at all.